backend/requirements.txt
asgiref==3.6.0
attrs==22.1.0
click==7.1.2
click-didyoumean==0.3.0
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.0.4
gunicorn==20.1.0
iniconfig==1.1.1
p==1.4.0
packaging==21.3
Pillow==9.4.0
pluggy==1.0.0
postgres==4.0
psycopg==3.1.8
psycopg2==2.9.5
psycopg2-binary==2.9.5
psycopg2-pool==1.1
PyJWT==2.6.0
pyparsing==3.0.9
pytest==7.2.0
python-dotenv==0.21.1
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
